+++ client/examples/output/build_alias_output.rs
@@ -49,7 +49,7 @@ async fn main() -> Result<()> {
         )))
         .finish_output(token_supply)?;
 
-    println!("{alias_output:?}");
+    println!("{alias_output:#?}");
 
     Ok(())
 }