+++ client/bindings/java/CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Changes from the Rust library;
+- `AliasOutputBuilderParams::stateMetadata` from `byte[]` to `String`;
 
 ## 1.0.0-rc.2 - 2023-02-09